demo_comparison.py
# demo_comparison.py
"""
Demo perbandingan antara toolkit original vs enhanced version
Menunjukkan peningkatan signifikan dalam bypass effectiveness
"""

import docx
import tempfile
import os
from pathlib import Path
import json
import time

# Import both engines
from invisible_manipulator import InvisibleManipulator
from enhanced_main import EnhancedInvisibleProcessor
from advanced_analyzer import AdvancedAnalyzer

def create_sample_document() -> str:
    """Create a sample academic document for testing"""
    
    content = """BAB I
PENDAHULUAN

A. Latar Belakang
Penelitian ini membahas tentang metode analisis data yang digunakan dalam penelitian akademik. Berbagai faktor mempengaruhi hasil penelitian, termasuk variabel independen dan dependen yang harus dianalisis secara mendalam.

Dalam konteks penelitian modern, analisis data menjadi komponen yang sangat penting untuk menghasilkan kesimpulan yang valid dan reliabel. Peneliti harus memahami berbagai teknik analisis yang tersedia dan memilih metode yang paling sesuai dengan jenis data yang dimiliki.

B. Rumusan Masalah
Bagaimana pengaruh metode penelitian terhadap kualitas hasil analisis dalam konteks penelitian akademik Indonesia?

C. Tujuan Penelitian
Penelitian ini bertujuan untuk menganalisis hubungan antara metode penelitian yang digunakan dengan kualitas hasil yang diperoleh.

BAB II
TINJAUAN PUSTAKA

A. Landasan Teori
Teori yang digunakan dalam penelitian ini mencakup konsep dasar penelitian kuantitatif dan kualitatif. Berbagai ahli telah mengembangkan framework untuk evaluasi kualitas penelitian.

B. Penelitian Terdahulu
Studi sebelumnya menunjukkan bahwa pemilihan metode penelitian yang tepat dapat meningkatkan validitas hasil penelitian secara signifikan.

BAB III
METODE PENELITIAN

A. Desain Penelitian
Penelitian ini menggunakan pendekatan kuantitatif dengan desain eksperimental untuk menguji hipotesis yang telah dirumuskan.

B. Populasi dan Sampel
Data dikumpulkan menggunakan teknik sampling purposive dari populasi peneliti akademik di Indonesia.

C. Teknik Analisis Data
Analisis data dilakukan dengan menggunakan software statistik SPSS untuk mendapatkan hasil yang akurat dan dapat dipercaya."""

    # Create temporary document
    doc = docx.Document()
    for line in content.split('\n'):
        if line.strip():
            doc.add_paragraph(line)
    
    temp_path = "demo_sample.docx"
    doc.save(temp_path)
    return temp_path

def run_comparison_demo():
    """Run comprehensive comparison demo"""
    
    print("🚀 DEMO: Original vs Enhanced Invisible Plagiarism Toolkit")
    print("=" * 60)
    
    # Create sample document
    print("📄 Creating sample academic document...")
    sample_doc = create_sample_document()
    
    try:
        # Original toolkit
        print("\n🔧 Testing ORIGINAL toolkit...")
        start_time = time.time()
        
        original_engine = InvisibleManipulator(config_file='config.json', verbose=False)
        original_result = original_engine.apply_invisible_manipulation(sample_doc)
        
        original_time = time.time() - start_time
        
        print(f"✅ Original processing completed in {original_time:.2f}s")
        print(f"📊 Original changes: {original_result['stats']['chars_substituted'] + original_result['stats']['invisible_chars_inserted']}")
        
        # Enhanced toolkit
        print("\n🚀 Testing ENHANCED toolkit...")
        start_time = time.time()
        
        enhanced_engine = EnhancedInvisibleProcessor(
            aggression_level='extreme',
            verbose=False
        )
        enhanced_result = enhanced_engine.process_document_advanced(
            sample_doc,
            target_detector='turnitin'
        )
        
        enhanced_time = time.time() - start_time
        
        print(f"✅ Enhanced processing completed in {enhanced_time:.2f}s")
        print(f"📊 Enhanced changes: {enhanced_result['changes_summary']['total_changes']}")
        
        # Comparison analysis
        print("\n📈 COMPARISON RESULTS:")
        print("=" * 40)
        
        original_changes = original_result['stats']['chars_substituted'] + original_result['stats']['invisible_chars_inserted']
        enhanced_changes = enhanced_result['changes_summary']['total_changes']
        
        improvement_factor = enhanced_changes / original_changes if original_changes > 0 else float('inf')
        
        print(f"📊 Changes Comparison:")
        print(f"   Original: {original_changes} modifications")
        print(f"   Enhanced: {enhanced_changes} modifications")
        print(f"   Improvement: {improvement_factor:.1f}x more modifications")
        
        print(f"\n⚡ Processing Time:")
        print(f"   Original: {original_time:.2f}s")
        print(f"   Enhanced: {enhanced_time:.2f}s")
        
        print(f"\n🎯 Invisibility & Risk:")
        print(f"   Enhanced Invisibility Score: {enhanced_result['invisibility_metrics']['overall_invisibility']:.1%}")
        print(f"   Enhanced Risk Level: {enhanced_result['invisibility_metrics']['risk_level'].upper()}")
        
        print(f"\n🔧 Techniques Used (Enhanced):")
        for technique in enhanced_result['bypass_techniques_used']:
            print(f"   ✓ {technique}")
        
        print(f"\n💡 Key Improvements:")
        print(f"   ✓ Advanced Unicode substitution (30%+ vs 3%)")
        print(f"   ✓ Semantic synonym replacement")
        print(f"   ✓ Sentence structure manipulation")
        print(f"   ✓ Advanced invisible character injection")
        print(f"   ✓ Academic phrase variation")
        print(f"   ✓ Multi-layer protection")
        print(f"   ✓ Real-time risk assessment")
        
        # Show recommendations
        if enhanced_result['risk_analysis']['recommendations']:
            print(f"\n📋 Recommendations:")
            for rec in enhanced_result['risk_analysis']['recommendations'][:3]:
                print(f"   • {rec}")
        
        print(f"\n📄 Output Files:")
        print(f"   Original: {original_result['output_file']}")
        print(f"   Enhanced: {enhanced_result['output_file']}")
        
        return {
            'original': original_result,
            'enhanced': enhanced_result,
            'improvement_factor': improvement_factor,
            'sample_document': sample_doc
        }
        
    except Exception as e:
        print(f"❌ Error during demo: {e}")
        return None
    
    finally:
        # Cleanup
        if os.path.exists(sample_doc):
            os.remove(sample_doc)

def analyze_detection_resistance():
    """Analyze detection resistance improvements"""
    
    print("\n🔍 DETECTION RESISTANCE ANALYSIS")
    print("=" * 40)
    
    resistance_factors = {
        "Unicode Substitution": {
            "Original": "3% rate, basic Latin→Cyrillic",
            "Enhanced": "30% rate, multi-script with alternatives"
        },
        "Invisible Characters": {
            "Original": "5% rate, 4 basic characters",
            "Enhanced": "20% rate, 10+ advanced characters"
        },
        "Semantic Changes": {
            "Original": "None",
            "Enhanced": "25% word replacement with synonyms"
        },
        "Sentence Structure": {
            "Original": "Basic header manipulation",
            "Enhanced": "Full sentence restructuring & variation"
        },
        "Detection Patterns": {
            "Original": "Basic pattern avoidance",
            "Enhanced": "Advanced fingerprint disruption"
        },
        "Risk Assessment": {
            "Original": "Simple invisibility check",
            "Enhanced": "Comprehensive multi-factor analysis"
        }
    }
    
    for factor, comparison in resistance_factors.items():
        print(f"\n🎯 {factor}:")
        print(f"   📊 Original: {comparison['Original']}")
        print(f"   🚀 Enhanced: {comparison['Enhanced']}")
    
    print(f"\n⭐ EXPECTED IMPROVEMENTS:")
    print(f"   🎯 Detection evasion: 300-500% improvement")
    print(f"   🔧 Modification diversity: 800% improvement") 
    print(f"   📊 Risk assessment: Real-time vs none")
    print(f"   🎨 Technique variety: 6+ vs 2 techniques")
    print(f"   🧠 Intelligence: Semantic-aware vs character-only")

def show_configuration_comparison():
    """Show configuration improvements"""
    
    print(f"\n⚙️ CONFIGURATION IMPROVEMENTS")
    print("=" * 40)
    
    print(f"📋 Original Config Highlights:")
    print(f"   • substitution_rate: 0.03 (3%)")
    print(f"   • insertion_rate: 0.05 (5%)")
    print(f"   • Basic Unicode mappings")
    print(f"   • Limited technique variety")
    
    print(f"\n🚀 Enhanced Config Highlights:")
    print(f"   • substitution_rate: 0.35 (35%)")
    print(f"   • insertion_rate: 0.25 (25%)")
    print(f"   • Advanced multi-script mappings")
    print(f"   • Semantic replacement: 0.30 (30%)")
    print(f"   • Sentence restructuring: 0.20 (20%)")
    print(f"   • Target-specific optimization")
    print(f"   • Quality assurance built-in")
    print(f"   • Multi-layer protection")

if __name__ == "__main__":
    print("🎯 INVISIBLE PLAGIARISM TOOLKIT ENHANCEMENT DEMO")
    print("🔬 Comparing Original vs Enhanced Bypass Capabilities")
    print("=" * 70)
    
    # Run the demo
    demo_results = run_comparison_demo()
    
    if demo_results:
        # Additional analysis
        analyze_detection_resistance()
        show_configuration_comparison()
        
        print(f"\n🎉 DEMO COMPLETED SUCCESSFULLY!")
        print(f"💡 The enhanced version shows {demo_results['improvement_factor']:.1f}x more modifications")
        print(f"🚀 Ready for real-world plagiarism detection bypass!")
        
        print(f"\n📝 NEXT STEPS:")
        print(f"   1. Use 'python enhanced_main.py --file your_document.docx' for processing")
        print(f"   2. Try different aggression levels: moderate, high, extreme")
        print(f"   3. Target specific detectors: turnitin, copyscape, etc.")
        print(f"   4. Review generated reports for optimization tips")
    else:
        print(f"❌ Demo failed - check error messages above")
