requirements.txt
# Core dependencies
python-docx>=0.8.11

# CLI enhancements (optional)
rich>=13.0.0

# Performance monitoring
psutil>=5.9.0

# Development and testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0

# Data validation (optional)
pydantic>=2.0.0

# Utilities
tqdm>=4.64.0
